services/rag_core/requirements.txt
fastapi
uvicorn
langchain
langchain-text-splitters
sentence-transformers
faiss-cpu
groq
requests
numpy
python-multipart
psycopg2-binary